backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pgvector==0.3.6
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
torch==2.5.0
numpy==2.1.2
pydantic==2.10.0
pulp==2.9.0
PyPDF2==3.0.1
google-genai>=1.0.0
PyJWT==2.8.0